Implemented language  sync with Gradio UI and removed dead language detection code
