+++ .env.example
@@ -19,7 +19,7 @@ GIGACHAT_BASE_URL=https://gigachat.sberdevices.ru/v1
 GIGACHAT_TIMEOUT=30
 AGENT_PROVIDER=mistral
 ENABLE_VECTOR_SIMILARITY=false
-CMW_DEFAULT_LANGUAGE=ru
+GRADIO_DEFAULT_LANGUAGE=ru
 CMW_DEFAULT_PORT=7860
 CMW_DEBUG_MODE=false
 CMW_VERBOSE_LOGGING=false
