+++ agent_ng/app_ng_modular.py
@@ -87,10 +87,9 @@ except ImportError as e1:
         _logger.error("Please check: 1) requirements_ng.txt installed 2) PYTHONPATH 3) circular imports 4) modules exist 5) working directory")
         raise ImportError(f"Failed to import required modules. Absolute: {e1}, Relative: {e2}")
 
-
 class NextGenApp:
     """LangChain-native Gradio application with modular tab architecture and i18n support"""
-    
+
     def __init__(self, language: str = "en"):
         # No global agent - only session-specific agents
         self.llm_manager = get_llm_manager()
@@ -102,7 +101,7 @@ class NextGenApp:
         self.language = language
         # Session-aware turn snapshots for analytics/logs (non-persistent)
         self.session_turn_snapshots = {}  # session_id -> turn_snapshot
-        
+
         # Initialize concurrency management
         try:
             from .concurrency_config import get_concurrency_config
@@ -115,7 +114,7 @@ class NextGenApp:
             from queue_manager import create_queue_manager
             self.concurrency_config = get_concurrency_config()
             self.queue_manager = create_queue_manager(self.concurrency_config)
-        
+
         # Session Management - Clean modular approach
         try:
             from .session_manager import SessionManager
@@ -123,14 +122,14 @@ class NextGenApp:
             # Fallback for when running as script
             from session_manager import SessionManager
         self.session_manager = SessionManager(language)
-        
+
         # Create i18n instance for the specified language
         self.i18n = create_i18n_instance(language)
-        
+
         # Initialize debug system - use a default session that will be accessible
         self.debug_streamer = get_debug_streamer("default")
         self.log_handler = get_log_handler("default")
-        
+
         # Initialize session-aware logging
         try:
             from .debug_streamer import set_session_context
@@ -139,7 +138,7 @@ class NextGenApp:
             from debug_streamer import set_session_context
         self.set_session_context = set_session_context
         # self.chat_interface = get_chat_interface("app_ng")  # Dead code - never used
-        
+
         # Initialize UI manager with i18n support
         try:
             self.ui_manager = get_ui_manager(language=language, i18n_instance=self.i18n)
@@ -148,35 +147,35 @@ class NextGenApp:
         except Exception as e:
             _logger.exception("Failed to initialize UI Manager: %s", e)
             self.ui_manager = None
-        
+
         # Initialize tab modules
         self.tabs = {}
         self.tab_instances = {}  # Store tab instances for event handlers
         self.components = {}
-        
+
         # Progress status storage with translation
         # Progress status is now managed per-session through session manager
         self.is_processing = False
         self.current_global_progress = get_translation_key("progress_ready", language)  # Global progress for timer updates
         self.last_progress_display = ""  # Cache last display to avoid unnecessary updates
-        
+
         # Persistent background asyncio loop for streaming to avoid closing gRPC/Gemini loop between turns
         self._stream_loop = None
         self._stream_loop_thread = None
         self._stream_queue_timeout_s = 0.1
         self._ensure_stream_loop()
-        
+
         # Initialize synchronously first, then start async initialization
         self._start_async_initialization()
-    
+
     def get_user_session_id(self, request: gr.Request = None) -> str:
         """Get session ID using clean session manager"""
         return self.session_manager.get_session_id(request)
-    
+
     def get_user_agent(self, session_id: str) -> NextGenAgent:
         """Get agent instance using clean session manager"""
         return self.session_manager.get_agent(session_id)
-    
+
     def _start_async_initialization(self):
         """Start async initialization in a new event loop"""
         try:
@@ -189,10 +188,10 @@ class NextGenApp:
             import threading
             def run_async_init():
                 asyncio.run(self._initialize_agent())
-            
+
             thread = threading.Thread(target=run_async_init, daemon=True)
             thread.start()
-    
+
     def _ensure_stream_loop(self):
         """Ensure a persistent background event loop exists and is running."""
         try:
@@ -202,22 +201,22 @@ class NextGenApp:
             # Recreate loop if previous loop errored out
             self._stream_loop = None
             self._stream_loop_thread = None
-        
+
         import threading
         import asyncio
         self._stream_loop = asyncio.new_event_loop()
-        
+
         def _run_loop_forever():
             asyncio.set_event_loop(self._stream_loop)
             self._stream_loop.run_forever()
-        
+
         self._stream_loop_thread = threading.Thread(target=_run_loop_forever, daemon=True)
         self._stream_loop_thread.start()
-    
+
     def _submit_stream_task(self, message, history, request, out_queue):
         """Submit the async streaming producer to the background loop and feed results into out_queue."""
         import asyncio
-        
+
         async def _producer():
             try:
                 async for result in self.stream_chat_with_agent(message, history, request):
@@ -228,26 +227,25 @@ class NextGenApp:
             finally:
                 # Sentinel to indicate completion
                 out_queue.put(StopIteration)
-        
+
         return asyncio.run_coroutine_threadsafe(_producer(), self._stream_loop)
-    
+
     async def _initialize_agent(self):
         """Initialize the session manager (no global agent needed)"""
         self.is_initializing = True
         self.debug_streamer.info("Starting session manager initialization", LogCategory.INIT)
         self.initialization_logs.append("🚀 " + get_translation_key("logs_initializing", self.language))
-        
+
         try:
             # Initialize session manager (creates agents on-demand per session)
             self.debug_streamer.info("Session manager ready", LogCategory.INIT)
-            
+
             self.initialization_logs.append("✅ " + get_translation_key("session_manager_ready", self.language))
             self.initialization_complete = True
-            
-            
+
             # Trigger UI update after initialization
             self._trigger_ui_update()
-                
+
         except Exception as e:
             try:
                 self.debug_streamer.error(f"Initialization failed: {str(e)}", LogCategory.INIT)
@@ -255,34 +253,34 @@ class NextGenApp:
                 # If debug streamer fails, just continue
                 pass
             self.initialization_logs.append(format_translation("error_initialization_failed", self.language, error=str(e)))
-        
+
         self.is_initializing = False
-    
+
     def get_conversation_summary(self, session_id: str, markdown: bool = False) -> str:
         """
         Generate unified conversation summary for logs and download frontmatter.
-        
+
         Args:
             session_id: Session ID to get summary for
             markdown: If True, format for markdown (download). If False, format for plain text (logs).
-            
+
         Returns:
             Formatted conversation summary string
         """
         try:
             from datetime import datetime
-            
+
             # Get session-specific agent and stats
             agent = self.session_manager.get_session_agent(session_id)
             if not agent:
                 self.debug_streamer.debug(f"No agent found for session: {session_id}")
                 return ""
-            
+
             # Get conversation stats
             stats = agent.get_stats()
             conversation_stats = stats.get("conversation_stats", {})
             llm_info = stats.get("llm_info", {})
-            
+
             # Get tool information from session-aware turn snapshot
             tool_names = set()
             tool_calls_count = 0
@@ -293,19 +291,19 @@ class NextGenApp:
                         tool_names.add(msg['name'])
                     elif msg.get('role') == 'assistant' and msg.get('tool_calls'):
                         tool_calls_count += len(msg['tool_calls'])
-            
+
             # Get roles sequence from snapshot
             roles_seq = []
             if snapshot:
                 roles_seq = [m.get("role", "?") for m in snapshot]
-            
+
             # Get provider/model info
             provider = llm_info.get('provider', 'unknown')
             model = llm_info.get('model', 'unknown')
-            
+
             # Format timestamp
             timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-            
+
             # Format tool information with total count
             if tool_names:
                 tool_counts = {}
@@ -318,7 +316,7 @@ class NextGenApp:
             else:
                 tools_text = "None"
                 total_tools = 0
-            
+
             # Get all used provider/model combinations from conversation history
             provider_models = set()
             if agent and hasattr(agent, "memory_manager"):
@@ -332,11 +330,11 @@ class NextGenApp:
                                 provider = msg.additional_kwargs.get('llm_provider', 'unknown')
                                 model = msg.additional_kwargs.get('llm_model', 'unknown')
                                 provider_models.add(f"{provider} / {model}")
-            
+
             # If no provider/model info found in messages, use current
             if not provider_models:
                 provider_models.add(f"{provider} / {model}")
-            
+
             # Format provider/model information
             if len(provider_models) == 1:
                 providers_text = list(provider_models)[0]
@@ -344,7 +342,7 @@ class NextGenApp:
             else:
                 providers_text = ", ".join(sorted(provider_models))
                 total_models = len(provider_models)
-            
+
             # Build summary based on format
             if markdown:
                 # Markdown format for download
@@ -361,28 +359,33 @@ class NextGenApp:
                 summary += f"{self._get_translation('roles_sequence')}: {' → '.join(roles_seq)}\n"
                 summary += f"{self._get_translation('tools_used_total')} ({total_tools}): {tools_text}\n"
                 summary += f"{self._get_translation('providers_models_total')} ({total_models}): {providers_text}\n"
-                summary += f"Total conversation tokens: {total_tokens:,}\n"
-                summary += f"Average tokens per conversation: {avg_tokens:,.0f}\n"
-            
+                # Add token statistics if available
+                total_tokens = conversation_stats.get('total_tokens', 0)
+                avg_tokens = conversation_stats.get('avg_tokens', 0)
+                if total_tokens > 0:
+                    summary += f"Total conversation tokens: {total_tokens:,}\n"
+                if avg_tokens > 0:
+                    summary += f"Average tokens per conversation: {avg_tokens:,.0f}\n"
+
             # Add newline for logs format
             if not markdown:
                 summary += "\n"
-            
+
             return summary
-            
+
         except Exception as e:
             self.debug_streamer.warning(f"Failed to generate conversation summary: {e}")
             return ""
-    
+
     def _get_translation(self, key: str) -> str:
         """Get translation for a key using the current language"""
         from .i18n_translations import get_translation_key
         return get_translation_key(key, self.language)
-    
+
     def is_ready(self) -> bool:
         """Check if the app is ready (session-based pattern)"""
         return self.initialization_complete and self.session_manager is not None
-    
+
     def clear_conversation(self, request: gr.Request = None) -> Tuple[List[Dict[str, str]], str]:
         """Clear the conversation history (LangChain-native pattern) - now properly session-aware"""
         if request:
@@ -400,11 +403,11 @@ class NextGenApp:
                     if hasattr(session_data.agent, 'token_tracker'):
                         session_data.agent.token_tracker.start_new_conversation()
                     self.debug_streamer.info("Conversation cleared (default session)")
-        
+
         # Reset progress status for all sessions
         # Progress status is now managed per-session through session manager
         return [], ""
-    
+
     def get_progress_status(self, request: gr.Request = None) -> str:
         """Get the current progress status for the UI - now properly session-aware"""
         if request:
@@ -412,7 +415,7 @@ class NextGenApp:
             return self.session_manager.get_status(session_id)
         # Fallback to default status if no request available
         return get_translation_key("progress_ready", self.language)
-    
+
     def update_progress_display(self, request: gr.Request = None) -> str:
         """Update progress display - session-aware, minimal with caching"""
         if not request:
@@ -422,30 +425,29 @@ class NextGenApp:
                 result = f"{icon} {self.current_global_progress}"
             else:
                 result = self.current_global_progress
-            
+
             # Only update if content changed to reduce UI blocking
             if result != self.last_progress_display:
                 self.last_progress_display = result
                 return result
             return self.last_progress_display
-        
+
         session_id = self.session_manager.get_session_id(request)
         status = self.session_manager.get_status(session_id)
-        
+
         # Add rotating clock icon if processing (UI handles rotation)
         if self.is_processing:
             icon = self.session_manager.get_clock_icon()
             result = f"{icon} {status}"
         else:
             result = status
-        
+
         # Only update if content changed to reduce UI blocking
         if result != self.last_progress_display:
             self.last_progress_display = result
             return result
         return self.last_progress_display
-    
-    
+
     def start_processing(self):
         """Mark that processing has started"""
         self.is_processing = True
@@ -453,7 +455,7 @@ class NextGenApp:
         self.current_global_progress = get_translation_key("progress_processing", self.language)
         # Reset cache to force update
         self.last_progress_display = ""
-    
+
     def stop_processing(self):
         """Mark that processing has stopped"""
         self.is_processing = False
@@ -461,7 +463,7 @@ class NextGenApp:
         self.current_global_progress = get_translation_key("progress_ready", self.language)
         # Reset cache to force update
         self.last_progress_display = ""
-    
+
     def get_conversation_history(self, session_id: str = "default") -> List[BaseMessage]:
         """Get the current conversation history (session-based pattern)"""
         if hasattr(self, 'session_manager'):
@@ -469,17 +471,16 @@ class NextGenApp:
             if session_data and session_data.agent:
                 return session_data.agent.get_conversation_history(session_id)
         return []
-    
-    
+
     async def stream_chat_with_agent(self, message: str, history: List[Dict[str, str]], request: gr.Request = None) -> AsyncGenerator[Tuple[List[Dict[str, str]], str], None]:
         """
         Stream chat with the agent using LangChain-native streaming patterns.
-        
+
         Args:
             message: User message
             history: Chat history
             progress: Optional Gradio Progress tracker
-            
+
         Yields:
             Updated history and empty message
         """
@@ -489,36 +490,36 @@ class NextGenApp:
             history.append({"role": "assistant", "content": error_msg})
             yield history, ""
             return
-        
+
         if not message.strip():
             yield history, ""
             return
-        
+
         # Track execution time
         start_time = time.time()
-        
+
         try:
             # Extract session ID from Gradio request for user isolation
             session_id = self.get_user_session_id(request)
             user_agent = self.get_user_agent(session_id)
-            
+
             # Set session context for logging
             self.set_session_context(session_id)
-            
+
             # Debug: Check which LLM instance is being used
             if user_agent and hasattr(user_agent, 'llm_instance') and user_agent.llm_instance:
                 llm_info = user_agent.get_llm_info()
                 print(f"🔍 DEBUG: Using session agent with LLM: {llm_info.get('provider', 'unknown')}/{llm_info.get('model_name', 'unknown')}")
             else:
                 print(f"❌ DEBUG: Session agent has no LLM instance!")
-            
+
             self.debug_streamer.info(f"Streaming message for session {session_id}: {message[:50]}...")
-            
+
             # Initialize response
-            
+
             # Add user message to history
             working_history = history + [{"role": "user", "content": message}, {"role": "assistant", "content": ""}]
-            
+
             # Get prompt token count for user message (will be displayed below assistant response)
             prompt_tokens = None
             if user_agent:
@@ -526,26 +527,26 @@ class NextGenApp:
                     prompt_tokens = user_agent.count_prompt_tokens_for_chat(history, message)
                 except Exception as e:
                     self.debug_streamer.warning(f"Failed to get prompt token count: {e}")
-            
+
             yield working_history, ""
-            
+
             # Stream response using simple streaming
             response_content = ""  # Initialize as empty string to prevent None concatenation
             tool_usage = ""
             assistant_message_index = -1  # Track the index of the assistant message in working_history
             # Tool messages are now added immediately to working_history during streaming
             streaming_error_handled = False  # Flag to track if streaming error was handled
-            
+
             # print(f"🔍 DEBUG: Starting streaming for session {session_id}")
             # print(f"🔍 DEBUG: Working history length before streaming: {len(working_history)}")
-            
+
             try:
                 async for event in user_agent.stream_message(message, session_id):
                     # Safety check for None event
                     if event is None:
                         # print("🔍 DEBUG: Received None event, skipping...")
                         continue
-                    
+
                     try:
                         event_type = event.get("type", "unknown")
                         content = event.get("content", "")
@@ -561,7 +562,7 @@ class NextGenApp:
                         streaming_error_handled = True
                         # print(f"🔍 DEBUG: Set streaming_error_handled to True in event processing")
                         continue
-                    
+
                     if event_type == "thinking":
                         # Agent is thinking - update or create assistant message
                         if assistant_message_index >= 0 and assistant_message_index < len(working_history):
@@ -572,7 +573,7 @@ class NextGenApp:
                             working_history.append({"role": "assistant", "content": content})
                             assistant_message_index = len(working_history) - 1
                         yield working_history, ""
-                        
+
                     elif event_type == "iteration_progress":
                         # Iteration progress - update progress display in sidebar
                         # Store progress status for UI update - session-specific
@@ -582,7 +583,7 @@ class NextGenApp:
                         # Reset cache to force update
                         self.last_progress_display = ""
                         yield working_history, ""
-                        
+
                     elif event_type == "completion":
                         # Final completion message - update progress display
                         self.session_manager.set_status(session_id, content)
@@ -600,12 +601,12 @@ class NextGenApp:
                             self.session_turn_snapshots[conversation_id] = ordered
                             self.debug_streamer.info(f"Turn snapshot stored for session {conversation_id} with {len(ordered)} messages")
                         yield working_history, ""
-                        
+
                     elif event_type == "tool_start":
                         # Tool is starting - immediately add to working history
                         tool_name = metadata.get("tool_name", "unknown") if metadata else "unknown"
                         tool_title = metadata.get("title", format_translation("tool_called", self.language, tool_name=tool_name)) if metadata else format_translation("tool_called", self.language, tool_name="unknown")
-                        
+
                         # Create tool message and immediately add to working history
                         tool_message = {
                             "role": "assistant", 
@@ -614,12 +615,12 @@ class NextGenApp:
                         }
                         working_history.append(tool_message)
                         yield working_history, ""
-                        
+
                     elif event_type == "tool_end":
                         # Tool completed - immediately add to working history
                         tool_name = metadata.get("tool_name", "unknown") if metadata else "unknown"
                         tool_title = metadata.get("title", format_translation("tool_called", self.language, tool_name=tool_name)) if metadata else format_translation("tool_called", self.language, tool_name="unknown")
-                        
+
                         # Create tool message and immediately add to working history
                         tool_message = {
                             "role": "assistant", 
@@ -628,12 +629,12 @@ class NextGenApp:
                         }
                         working_history.append(tool_message)
                         yield working_history, ""
-                        
+
                     elif event_type == "content":
                         # Stream content from response - ensure content is not None
                         content_to_add = safe_string(content)
                         # print(f"🔍 DEBUG: Content event - content: '{content_to_add}', length: {len(content_to_add)}")
-                        
+
                         # Only add line break when LLM starts answering after tool messages
                         if assistant_message_index < 0 and response_content == "":
                             # This is the very first content, check if there are recent tool messages
@@ -642,13 +643,13 @@ class NextGenApp:
                                 if i >= 0 and working_history[i] and working_history[i].get("metadata", {}).get("title"):
                                     has_recent_tool_messages = True
                                     break
-                            
+
                             if has_recent_tool_messages:
                                 # Add lean line break only when LLM starts answering after tools
                                 content_to_add = "\n" + content_to_add
-                        
+
                         response_content += content_to_add
-                        
+
                         # Update or create assistant message
                         if assistant_message_index >= 0 and assistant_message_index < len(working_history):
                             # Update existing assistant message
@@ -659,23 +660,23 @@ class NextGenApp:
                             working_history.append({"role": "assistant", "content": response_content})
                             assistant_message_index = len(working_history) - 1
                             # print(f"🔍 DEBUG: Created new assistant message at index {assistant_message_index}")
-                        
+
                         yield working_history, ""
-                        
+
                     elif event_type == "error":
                         # Add error message to response content so users can see what went wrong
                         response_content += content + "\n\n"
-                        
+
                         # Update assistant message with error
                         if assistant_message_index >= 0 and assistant_message_index < len(working_history):
                             working_history[assistant_message_index] = {"role": "assistant", "content": response_content}
                         else:
                             working_history.append({"role": "assistant", "content": response_content})
                             assistant_message_index = len(working_history) - 1
-                        
+
                         streaming_error_handled = True
                         yield working_history, ""
-            
+
             except Exception as e:
                 import traceback
                 _logger.error("Error streaming message: %s", e, exc_info=True)
@@ -683,15 +684,15 @@ class NextGenApp:
                 # print(f"🔍 DEBUG: Set streaming_error_handled to True in streaming loop")
                 # Continue with the rest of the processing even if streaming fails
                 pass
-            
+
             # Add API token count to final response
             # Add token counts below assistant response
             token_displays = []
-            
+
             # Add prompt tokens if available
             if prompt_tokens:
                 token_displays.append(format_translation("prompt_tokens", self.language, tokens=prompt_tokens.formatted))
-            
+
             # Add API tokens if available from session-specific agent
             if user_agent:
                 try:
@@ -707,7 +708,7 @@ class NextGenApp:
                 except Exception as e:
                     # print(f"🔍 DEBUG: API token error: {e}")
                     self.debug_streamer.warning(f"Failed to get API token count: {e}")
-            
+
             # Add provider/model information if available - use session-specific agent
             if user_agent and hasattr(user_agent, 'get_llm_info'):
                 try:
@@ -721,10 +722,10 @@ class NextGenApp:
                 except Exception as e:
                     # print(f"🔍 DEBUG: Provider/model display error: {e}")
                     self.debug_streamer.warning(f"Failed to get provider/model info: {e}")
-            
+
             # Calculate execution time for the entire response
             execution_time = time.time() - start_time
-            
+
             # Add deduplication stats if available from session-specific agent
             if user_agent and hasattr(user_agent, '_deduplication_stats'):
                 dedup_stats = user_agent._deduplication_stats.get(session_id, {})
@@ -732,13 +733,13 @@ class NextGenApp:
                     dedup_summary = []
                     total_duplicates = 0
                     total_tool_calls = 0
-                    
+
                     for tool_key, stats in dedup_stats.items():
                         total_tool_calls += stats['total_calls']
                         if stats['duplicates'] > 0:
                             dedup_summary.append(f"{stats['tool_name']}: {stats['duplicates']}")
                             total_duplicates += stats['duplicates']
-                    
+
                     if dedup_summary:
                         # Show per-tool breakdown
                         per_tool_breakdown = ", ".join(dedup_summary)
@@ -746,12 +747,12 @@ class NextGenApp:
                                                                duplicates=total_duplicates, 
                                                                breakdown=per_tool_breakdown))
                         # print(f"🔍 DEBUG: Added deduplication stats: {total_duplicates} duplicates")
-                    
+
                     # Add total tool calls count
                     if total_tool_calls > 0:
                         token_displays.append(format_translation("total_tool_calls", self.language, calls=total_tool_calls))
                         # print(f"🔍 DEBUG: Added total tool calls: {total_tool_calls}")
-            
+
             # Add token statistics as a separate metadata block
             if token_displays:
                 # Add execution time to the token display
@@ -765,7 +766,7 @@ class NextGenApp:
                 }
                 working_history.append(token_metadata_message)
                 # print(f"🔍 DEBUG: Added token metadata block: {token_display}")
-            
+
             # Tool messages are now added immediately during streaming, no need to add them here
             # Ensure tool messages are preserved and not overwritten
             # print(f"🔍 DEBUG: Final working history length: {len(working_history)}")
@@ -785,15 +786,15 @@ class NextGenApp:
                 elif msg and msg.get("role") == "assistant":
                     # print(f"🔍 DEBUG: Assistant message {i}: {len(msg.get('content', ''))} chars")
                     pass
-            
+
             # Stop processing state
             self.stop_processing()
-            
+
             # Final yield with updated stats
             # print(f"🔍 DEBUG: Final yield - working_history length: {len(working_history)}")
             # print(f"🔍 DEBUG: Final yield - response_content length: {len(response_content)}")
             yield working_history, ""
-            
+
         except Exception as e:
             # Log error to terminal but don't add to chat response
             try:
@@ -807,14 +808,14 @@ class NextGenApp:
             # Continue gracefully - don't stop processing for streaming errors
             # The response might still be valid even with streaming issues
             yield working_history, ""
-    
+
     def _create_event_handlers(self) -> Dict[str, Any]:
         """Create event handlers for all tabs"""
         handlers = {
             # Chat handlers (core functionality)
             "stream_message": self._stream_message_wrapper,
             "clear_chat": self.clear_conversation,
-            
+
             # Status and monitoring handlers
             "update_status": self._update_status,
             "update_token_budget": self._update_token_budget,
@@ -824,30 +825,30 @@ class NextGenApp:
             "trigger_ui_update": self.trigger_ui_update,
             "get_progress_status": self.get_progress_status,
             "update_progress_display": self.update_progress_display,
-            
+
             # LLM selection handlers - removed auto-refresh handler
             # LLM selection components update only when explicitly triggered
         }
-        
+
         # Add language switch handler if this is the language detection app
-        
+
         return handlers
-    
+
     def _stream_message_wrapper(self, message: str, history: List[Dict[str, str]], request: gr.Request = None):
         """Stream a message to the agent (synchronous wrapper)"""
         if not message.strip():
             yield history, ""
             return
-        
+
         # Start processing state for icon rotation
         self.start_processing()
-        
+
         # Always use the persistent background loop to avoid closing gRPC/Gemini resources
         self._ensure_stream_loop()
         import queue
         out_queue = queue.Queue()
         future = self._submit_stream_task(message, history, request, out_queue)
-        
+
         try:
             while True:
                 item = out_queue.get()
@@ -861,54 +862,53 @@ class NextGenApp:
                     future.cancel()
             except Exception:
                 pass
-        
+
         # Refresh UI after streaming completes (EVENT-DRIVEN)
         self._refresh_ui_after_message()
-    
+
     def _update_status(self, request: gr.Request = None) -> str:
         """Update status display - always session-aware"""
         # Use stats tab for proper formatting (now always session-aware)
         stats_tab = self.tab_instances.get('stats')
         if stats_tab and hasattr(stats_tab, 'format_stats_display'):
             return stats_tab.format_stats_display(request)
-        
+
         # Final fallback
         if self.is_ready():
             return get_translation_key("agent_ready", self.language)
         else:
             return get_translation_key("agent_initializing", self.language)
-    
+
     def _update_token_budget(self, request: gr.Request = None) -> str:
         """Update token budget display - delegates to chat tab with session awareness"""
         chat_tab = self.tab_instances.get('chat')
         if chat_tab and hasattr(chat_tab, 'format_token_budget_display'):
             return chat_tab.format_token_budget_display(request)
-        
+
         # Fallback token budget
         return get_translation_key("token_budget_initializing", self.language)
-    
-    
+
     def _refresh_logs(self, request: gr.Request = None) -> str:
         """Refresh logs display - delegates to logs tab with session awareness"""
         logs_tab = self.tab_instances.get('logs')
         if logs_tab and hasattr(logs_tab, 'get_initialization_logs'):
             return logs_tab.get_initialization_logs(request)
-        
+
         # Fallback logs
         return "\n".join(self.initialization_logs) if self.initialization_logs else "No logs available"
-    
+
     def _refresh_stats(self, request: gr.Request = None) -> str:
         """Refresh stats display - delegates to stats tab with session awareness"""
         stats_tab = self.tab_instances.get('stats')
         if stats_tab and hasattr(stats_tab, 'format_stats_display'):
             return stats_tab.format_stats_display(request)
-        
+
         # Fallback stats
         if self.is_ready():
             return get_translation_key("agent_ready", self.language)
         else:
             return get_translation_key("agent_initializing", self.language)
-    
+
     def _trigger_ui_update(self):
         """Trigger UI update after agent initialization or message processing"""
         try:
@@ -918,21 +918,21 @@ class NextGenApp:
         except Exception as e:
             # print(f"🔍 DEBUG: Error triggering UI update: {e}")
             pass
-    
+
     def check_and_clear_ui_update(self) -> bool:
         """Check if UI update is needed and clear the flag"""
         if self._ui_update_needed:
             self._ui_update_needed = False
             return True
         return False
-    
+
     def update_all_ui_components(self) -> Tuple[str, str, str]:
         """Update all UI components and return their values"""
         status = self._update_status()
         stats = self._refresh_stats()
         logs = self._refresh_logs()
         return status, stats, logs
-    
+
     def trigger_ui_update(self):
         """Trigger UI update after agent initialization or message processing"""
         try:
@@ -943,28 +943,28 @@ class NextGenApp:
         except Exception as e:
             # print(f"🔍 DEBUG: Error triggering UI update: {e}")
             pass
-    
+
     def _refresh_ui_after_message(self):
         """Refresh all UI components after a message is processed (EVENT-DRIVEN)"""
         try:
             # Trigger UI update after message processing
             self.trigger_ui_update()
-            
+
             # print("🔍 DEBUG: UI refreshed after message completion (EVENT-DRIVEN)")
-            
+
         except Exception as e:
             # print(f"🔍 DEBUG: Error refreshing UI after message: {e}")
             pass
-    
+
     def create_interface(self) -> gr.Blocks:
         """Create the Gradio interface using UI Manager and modular tabs"""
         # Validate UI Manager
         if not self.ui_manager:
             raise RuntimeError("UI Manager not available - cannot create interface")
-        
+
         # Create event handlers
         event_handlers = self._create_event_handlers()
-        
+
         # Create tab modules with error handling
         tab_modules = []
         try:
@@ -975,7 +975,7 @@ class NextGenApp:
                 self.tab_instances['chat'] = chat_tab
             else:
                 _logger.warning("ChatTab not available")
-                
+
             if LogsTab:
                 logs_tab = LogsTab(event_handlers, language=self.language, i18n_instance=self.i18n)
                 logs_tab.set_main_app(self)  # Pass main app reference
@@ -983,7 +983,7 @@ class NextGenApp:
                 self.tab_instances['logs'] = logs_tab
             else:
                 _logger.warning("LogsTab not available")
-                
+
             if StatsTab:
                 stats_tab = StatsTab(event_handlers, language=self.language, i18n_instance=self.i18n)
                 stats_tab.set_main_app(self)  # Set reference to main app for session management
@@ -994,112 +994,64 @@ class NextGenApp:
         except Exception as e:
             _logger.exception("Error creating tab modules: %s", e)
             raise
-        
+
         # Use UI Manager to create interface
         try:
             demo = self.ui_manager.create_interface(tab_modules, event_handlers)
         except Exception as e:
             _logger.exception("Error creating interface: %s", e)
             raise
-        
+
         # Configure concurrency and queuing
         self.queue_manager.configure_queue(demo)
-        
+
         # Consolidate all components from UI Manager (single source of truth)
         self.components = self.ui_manager.get_components()
-        
-        # No global agent - tabs will get session-specific agents as needed
-        
-        return demo
 
+        # No global agent - tabs will get session-specific agents as needed
+
+        return demo
 
 class NextGenAppWithLanguageDetection(NextGenApp):
     """LangChain-native Gradio application with dynamic language detection and switching"""
-    
+
     def __init__(self, language: str = "en"):
         super().__init__(language)
         self.current_language = language
         self.supported_languages = ["en", "ru"]
-        
-        # Import the simple language detector
-        try:
-            from .simple_language_detector import get_current_language
-        except ImportError:
-            # Fallback for when running as script
-            from simple_language_detector import get_current_language
-        self.get_current_language = get_current_language
-    
-    
-    def detect_language_from_url(self):
-        """Detect language from URL parameters using environment variables or sys.argv"""
-        try:
-            # Check if we're running with Gradio and can access URL parameters
-            import os
-            import sys
-            
-            # Check for language parameter in command line arguments
-            for i, arg in enumerate(sys.argv):
-                if arg == '--lang' and i + 1 < len(sys.argv):
-                    lang = sys.argv[i + 1].lower()
-                    if lang in self.supported_languages:
-                        print(f"🌐 Language detected from command line: {lang}")
-                        return lang
-                elif arg.startswith('--lang='):
-                    lang = arg.split('=')[1].lower()
-                    if lang in self.supported_languages:
-                        print(f"🌐 Language detected from command line: {lang}")
-                        return lang
-            
-            # Check environment variable (can be set by Gradio)
-            lang_env = os.environ.get('GRADIO_LANG', '').lower()
-            if lang_env in self.supported_languages:
-                print(f"🌐 Language detected from environment: {lang_env}")
-                return lang_env
-            
-            # Check for URL parameter in environment (Gradio might set this)
-            url_lang = os.environ.get('LANG_PARAM', '').lower()
-            if url_lang in self.supported_languages:
-                print(f"🌐 Language detected from URL parameter: {url_lang}")
-                return url_lang
-            
-            return None
-            
-        except Exception as e:
-            print(f"⚠️ URL language detection failed: {e}")
-            return None
-    
-    def detect_language(self, request=None):
-        """Detect language from URL parameters, headers, or browser settings"""
-        # Default to Russian
-        detected_lang = "ru"
-        
-        try:
-            # Check URL parameters first
-            if request and hasattr(request, 'quick_params'):
-                lang_param = request.quick_params.get('lang', '').lower()
-                if lang_param in self.supported_languages:
-                    detected_lang = lang_param
-                    print(f"🌐 Language detected from URL parameter: {detected_lang}")
-                    return detected_lang
-            
-            # Check Accept-Language header
-            if request and hasattr(request, 'headers'):
-                accept_lang = request.headers.get('Accept-Language', '')
-                if 'ru' in accept_lang.lower():
-                    detected_lang = "ru"
-                    print(f"🌐 Language detected from browser: {detected_lang}")
-                    return detected_lang
-            
-            # Fallback to default
-            print(f"🌐 Using default language: {detected_lang}")
-            return detected_lang
-            
-        except Exception as e:
-            print(f"⚠️ Language detection failed: {e}, using default: {detected_lang}")
-            return detected_lang
-    
-    
 
+        # Language detection is now integrated directly into this class
+
+    def get_current_language(self, request: gr.Request = None) -> str:
+        """
+        Get current language using Gradio's I18n system with GRADIO_DEFAULT_LANGUAGE as primary source.
+
+        Args:
+            request: Gradio request object (optional)
+
+        Returns:
+            Language code ('en' or 'ru')
+        """
+        try:
+            # Primary: Use GRADIO_DEFAULT_LANGUAGE environment variable
+            import os
+            from dotenv import load_dotenv
+            load_dotenv()  # Load .env file
+
+            gradio_lang = os.getenv('GRADIO_DEFAULT_LANGUAGE', '').lower()
+            if gradio_lang is not None:
+                return gradio_lang
+
+            # Secondary: Use Gradio's I18n system for detection
+            i18n = gr.I18n(
+                en={"language": "en"},
+                ru={"language": "ru"}
+            )
+            detected_language = i18n("language")
+            if detected_language is not None:
+                return detected_language
+        except Exception:
+            return "ru"
 
 # Global demo variable for single port architecture
 demo = None
@@ -1107,15 +1059,22 @@ demo = None
 def get_demo_with_language_detection():
     """Get or create the demo interface with language detection support"""
     global demo
-    
+
     if demo is None:
         try:
-            # Create app with language detection capability
-            app = NextGenAppWithLanguageDetection()
+            # First, detect the language using the elegant i18n system
+            temp_app = NextGenAppWithLanguageDetection()
+            detected_language = temp_app.get_current_language()
+
+            # Create app with the detected language
+            app = NextGenAppWithLanguageDetection(language=detected_language)
             demo = app.create_interface()
+
             # Ensure the demo has the required attributes for Gradio reloading
             if not hasattr(demo, '_queue'):
                 demo._queue = None
+
+            _logger.info(f"🌐 Demo created with detected language: {detected_language}")
         except Exception as e:
             _logger.exception("Error creating demo: %s", e)
             # Create a minimal working demo with required attributes
@@ -1126,46 +1085,32 @@ def get_demo_with_language_detection():
             if not hasattr(fallback_demo, '_queue'):
                 fallback_demo._queue = None
             demo = fallback_demo
-    
+
     return demo
 
-def get_demo(language: str = "en"):
-    """Get or create the demo interface for the specified language (legacy support)"""
-    # For backward compatibility, create language-specific demo
-    try:
-        app = NextGenApp(language=language)
-        return app.create_interface()
-    except Exception as e:
-        _logger.exception("Error creating demo for language %s: %s", language, e)
-        # Create a minimal working demo to prevent KeyError
-        with gr.Blocks() as demo:
-            gr.Markdown("# CMW Platform Agent")
-            gr.Markdown("Application is initializing...")
-        return demo
-
-# Create a safe demo instance for Gradio reloading
-# This prevents the _queue attribute error by ensuring demo is always valid
-def create_safe_demo():
-    """Create a safe demo instance that won't cause reload errors"""
-    try:
-        demo_instance = get_demo("en")  # Default to English
-        # Ensure the demo has the required attributes for Gradio reloading
-        if not hasattr(demo_instance, '_queue'):
-            demo_instance._queue = None
-        return demo_instance
-    except Exception as e:
-        _logger.exception("Error creating safe demo: %s", e)
-        # Create a minimal working demo with required attributes
-        with gr.Blocks() as demo:
-            gr.Markdown("# CMW Platform Agent")
-            gr.Markdown("Application is initializing...")
-        # Ensure required attributes exist
-        if not hasattr(demo, '_queue'):
-            demo._queue = None
-        return demo
+def create_fallback_demo():
+    """Create a minimal fallback demo for error cases"""
+    with gr.Blocks() as fallback_demo:
+        gr.Markdown("# CMW Platform Agent")
+        gr.Markdown("Application is initializing...")
+    # Ensure required attributes exist
+    if not hasattr(fallback_demo, '_queue'):
+        fallback_demo._queue = None
+    return fallback_demo
 
 # Initialize demo for Gradio reloading - use language detection
-demo = get_demo_with_language_detection()
+# This ensures demo is available for Gradio's reload mechanism
+def initialize_demo():
+    """Initialize the demo with proper error handling"""
+    try:
+        return get_demo_with_language_detection()
+    except Exception as e:
+        _logger.exception("Failed to initialize demo: %s", e)
+        # Create a minimal fallback demo
+        return create_fallback_demo()
+
+# Initialize demo - this will be available for Gradio's static analysis
+demo = initialize_demo()
 
 def reload_demo():
     """Reload the demo for Gradio hot reloading"""
@@ -1187,7 +1132,7 @@ def reload_demo():
 def find_available_port(start_port=7860, max_attempts=10):
     """Find an available port starting from start_port"""
     import socket
-    
+
     for port in range(start_port, start_port + max_attempts):
         try:
             with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
@@ -1202,11 +1147,11 @@ def main():
     import argparse
     import sys
     import os
-    
+
     # Setup LangSmith environment first
     from agent_ng.langsmith_config import setup_langsmith_environment
     setup_langsmith_environment()
-    
+
     # Parse command line arguments
     parser = argparse.ArgumentParser(description='CMW Platform Agent')
     parser.add_argument('-en', '--english', action='store_true', help='Start in English')
@@ -1215,39 +1160,39 @@ def main():
     parser.add_argument('--auto-port', action='store_true', help='Automatically find an available port')
     parser.add_argument('--config', action='store_true', help='Show current configuration')
     args = parser.parse_args()
-    
+
     # Show configuration if requested
     if args.config:
         config.print_config()
         return
-    
+
     # Get language settings from central config
     language_settings = get_language_settings()
     port_settings = get_port_settings()
-    
+
     # Determine language from command line, environment variable, or config default
     # Priority: Command line > Environment variable > Config default
     language = language_settings['default_language']
-    
+
+    try:
+        # Create a temporary app instance to use the integrated language detection
+        temp_app = NextGenAppWithLanguageDetection()
+        detected_language = temp_app.get_current_language()
+        if detected_language in ["en", "ru"]:
+            language = detected_language
+            _logger.info(f"🌐 Language overridden with Gradio detection: {detected_language}")
+    except Exception as e:
+        _logger.warning(f"⚠️ Language detection failed: {e}, using dotenv language: {language}")
+
     if args.russian:
         language = "ru"
     elif args.english:
         language = "en"
-    
+
     # Override language with Gradio I18n detection if no command line override
-    if not (args.russian or args.english):
-        try:
-            from agent_ng.simple_language_detector import get_current_language
-            detected_language = get_current_language()
-            if detected_language in ["en", "ru"]:
-                language = detected_language
-                _logger.info(f"🌐 Language overridden with Gradio detection: {detected_language}")
-        except Exception as e:
-            _logger.warning(f"⚠️ Language detection failed: {e}, using dotenv language: {language}")
-    
     # Determine port
     default_port = args.port if args.port is not None else port_settings['default_port']
-    
+
     if args.auto_port:
         port = find_available_port(default_port, port_settings['auto_port_range'])
         if port is None:
@@ -1255,15 +1200,15 @@ def main():
             sys.exit(1)
     else:
         port = default_port
-    
+
     _logger.info("Starting LangChain-Native LLM Agent App with language detection...")
     _logger.info("Language: %s", language.upper())
     _logger.info("Port: %s", port)
-    
+
     # Create app with specified language
     app = NextGenAppWithLanguageDetection(language=language)
     demo = app.create_interface()
-    
+
     _logger.info("Launching Gradio interface on port %s with language switching...", port)
     demo.launch(
         debug=True,
@@ -1272,7 +1217,6 @@ def main():
         server_port=port,
         show_error=True
     )
-
 
 if __name__ == "__main__":
     main()
